+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.53.0] - 2021-03-12
+### Added
+- gRPC: accept keepalive parameters for gRPC outbound configuration.
+- observability: deprecation of logging levels `applicationError` and `serverError` in favor
+  of `serverError` and `clientError`.
+- The `transport.Request` provides a new field `CallerProcedure`. A header `rpc-caller-procedure`
+  is also propagated for outbound calls. This field represents the origin procedure where the
+  requests was sent.
+### Fixed
+- peer: concurrency issue when using `rand` for the peer strategies `randpeer` and `tworandomchoices`
+
 ## [1.52.0] - 2021-02-12
 ### Added
 - gRPC: accept keepalive parameters for the outbound gRPC connection.
@@ -1328,6 +1339,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.53.0]: https://github.com/yarpc/yarpc-go/compare/v1.52.0...v1.53.0
 [1.52.0]: https://github.com/yarpc/yarpc-go/compare/v1.51.0...v1.52.0
 [1.51.0]: https://github.com/yarpc/yarpc-go/compare/v1.50.0...v1.51.0
 [1.50.0]: https://github.com/yarpc/yarpc-go/compare/v1.49.1...v1.50.0